% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{.onLoad}
\alias{.onLoad}
\title{Package load hook for}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Character string giving the library directory.}

\item{pkgname}{Character string giving the package name.}
}
\description{
Internal initialization hook called when the package is loaded.
No heavy work (such as initializing Python or installing dependencies)
should be performed here.
}
\keyword{internal}
