% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{.rt_add}
\alias{.rt_add}
\title{Reticulate backend: add rows to table}
\usage{
.rt_add(tbl, data)
}
\arguments{
\item{tbl}{A \code{lancedb_table} object.}

\item{data}{Data to add.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Backend-specific implementation for adding data using the
Python LanceDB client.
}
\keyword{internal}
