% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{.rt_connect}
\alias{.rt_connect}
\title{Reticulate backend: connect to a LanceDB database}
\usage{
.rt_connect(uri)
}
\arguments{
\item{uri}{Character scalar giving the database location.}
}
\value{
A \code{lancedb_connection} object.
}
\description{
Backend-specific implementation of \code{lancedb_connect()} using
the Python LanceDB client.
}
\keyword{internal}
