% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{.rt_create_table}
\alias{.rt_create_table}
\title{Reticulate backend: create a table}
\usage{
.rt_create_table(con, name, data, mode)
}
\arguments{
\item{con}{A \code{lancedb_connection} object.}

\item{name}{Character scalar giving the table name.}

\item{data}{Data used to create the table.}

\item{mode}{Creation mode (\code{"overwrite"} or \code{"create"}).}
}
\value{
A \code{lancedb_table} object.
}
\description{
Backend-specific implementation of \code{lancedb_create_table()}
using the Python LanceDB client.
}
\keyword{internal}
