% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{.rt_open_table}
\alias{.rt_open_table}
\title{Reticulate backend: open an existing table}
\usage{
.rt_open_table(con, name)
}
\arguments{
\item{con}{A \code{lancedb_connection} object.}

\item{name}{Character scalar giving the table name.}
}
\value{
A \code{lancedb_table} object.
}
\description{
Backend-specific implementation of \code{lancedb_open_table()}
using the Python LanceDB client.
}
\keyword{internal}
