% Generated by roxygen2: do not edit by hand
\name{lancedb_checkout_latest}
\alias{lancedb_checkout_latest}
\title{Checkout the Latest Table Version}
\usage{
lancedb_checkout_latest(table)
}
\arguments{
\item{table}{A \code{lancedb_table} object.}
}
\value{
The table (invisibly).
}
\description{
Returns the table to its latest version after a previous
\code{\link{lancedb_checkout}} call.
}
\examples{
\dontrun{
lancedb_checkout_latest(tbl)
}
}
\seealso{
\code{\link{lancedb_checkout}}, \code{\link{lancedb_restore}}
}
