% Generated by roxygen2: do not edit by hand
\name{lancedb_cleanup_old_versions}
\alias{lancedb_cleanup_old_versions}
\title{Clean Up Old Table Versions}
\usage{
lancedb_cleanup_old_versions(table, older_than_days = 7L)
}
\arguments{
\item{table}{A \code{lancedb_table} object.}
\item{older_than_days}{Integer. Remove versions older than this many days.
Default 7.}
}
\value{
The table (invisibly).
}
\description{
Prunes old versions that are older than the specified number of days.
}
\examples{
\dontrun{
lancedb_cleanup_old_versions(tbl, older_than_days = 30)
}
}
\seealso{
\code{\link{lancedb_compact_files}}
}
