% Generated by roxygen2: do not edit by hand
\name{lancedb_fast_search}
\alias{lancedb_fast_search}
\title{Enable Fast Search Mode}
\usage{
lancedb_fast_search(.data)
}
\arguments{
\item{.data}{A \code{lancedb_lazy} object.}
}
\value{
A new \code{lancedb_lazy} object with fast search enabled.
}
\description{
Enables approximate (faster) search that may skip some results for speed.
Useful for large-scale searches where approximate results are acceptable.
}
\examples{
\dontrun{
lancedb_search(tbl, runif(128)) \%>\%
  lancedb_fast_search() \%>\%
  slice_head(n = 10) \%>\%
  collect()
}
}
\seealso{
\code{\link{lancedb_search}}, \code{\link{lancedb_postfilter}}
}
