% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{lancedb_open_table}
\alias{lancedb_open_table}
\title{Open an existing LanceDB table}
\usage{
lancedb_open_table(con, name)
}
\arguments{
\item{con}{A \code{lancedb_connection} object.}

\item{name}{Character scalar giving the table name.}
}
\value{
A \code{lancedb_table} object.
}
\description{
Opens an existing table within a LanceDB database connection.
This function does not create the table if it does not exist.
}
