% Generated by roxygen2: do not edit by hand
\name{lancedb_restore}
\alias{lancedb_restore}
\title{Restore a Checked-Out Version}
\usage{
lancedb_restore(table)
}
\arguments{
\item{table}{A \code{lancedb_table} object.}
}
\value{
The table (invisibly).
}
\description{
Creates a new version of the table from the currently checked-out version.
Must be called after \code{\link{lancedb_checkout}}.
}
\examples{
\dontrun{
lancedb_checkout(tbl, version = 1)
lancedb_restore(tbl)
}
}
\seealso{
\code{\link{lancedb_checkout}}, \code{\link{lancedb_checkout_latest}}
}
