% Generated by roxygen2: do not edit by hand
\name{lancedb_to_arrow}
\alias{lancedb_to_arrow}
\title{Convert Entire Table to Arrow}
\usage{
lancedb_to_arrow(table)
}
\arguments{
\item{table}{A \code{lancedb_table} object.}
}
\value{
An \code{arrow::Table}.
}
\description{
Reads the full table contents as an Arrow Table.
}
\examples{
\dontrun{
arrow_tbl <- lancedb_to_arrow(tbl)
}
}
