% Generated by roxygen2: do not edit by hand
\name{lancedb_version}
\alias{lancedb_version}
\title{Get Current Table Version}
\usage{
lancedb_version(table)
}
\arguments{
\item{table}{A \code{lancedb_table} object.}
}
\value{
An integer version number.
}
\description{
Returns the current version number of a LanceDB table.
}
\examples{
\dontrun{
lancedb_version(tbl)
}
}
