% Generated by roxygen2: do not edit by hand
\name{lancedb_with_row_id}
\alias{lancedb_with_row_id}
\title{Include Row IDs in Results}
\usage{
lancedb_with_row_id(.data)
}
\arguments{
\item{.data}{A \code{lancedb_lazy} object.}
}
\value{
A new \code{lancedb_lazy} object that will include row IDs.
}
\description{
Adds the internal LanceDB row ID (\code{_rowid}) column to the query results.
This is useful for identifying specific rows for updates or deletes.
}
\examples{
\dontrun{
lancedb_scan(tbl) \%>\%
  lancedb_with_row_id() \%>\%
  slice_head(n = 10) \%>\%
  collect()
}
}
\seealso{
\code{\link{lancedb_scan}}, \code{\link{lancedb_search}}
}
