% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{print.lancedb_info}
\alias{print.lancedb_info}
\title{Print method for lancedb_info objects}
\usage{
\method{print}{lancedb_info}(x, ...)
}
\arguments{
\item{x}{A \code{lancedb_info} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Nicely formats and prints information returned by
\code{lancedb_info()}.
}
\keyword{internal}
