% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{print.lancedbr_info}
\alias{print.lancedbr_info}
\title{Print method for lancedbr_info objects}
\usage{
\method{print}{lancedbr_info}(x, ...)
}
\arguments{
\item{x}{A \code{lancedbr_info} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Nicely formats and prints information returned by
\code{lancedbr_info()}.
}
\keyword{internal}
